Develop

+++ src/classification/server.py
@@ -2,7 +2,9 @@
 
 from flask import Flask
 from flask import request
+import time
 import json
+import logging
 import os
 
 import file_handler
@@ -13,13 +15,16 @@ import logistic_regression_classification as classification
 
 app = Flask(__name__)
 
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
 FILE_ENCODING = 'utf-8'
 
 if 'LABELS_FILE' not in os.environ:
-    print("WARNING: LABELS_FILE not found in environment variables!")
+    logger.error("LABELS_FILE not found in environment variables!")
 
 if 'PREDS_FILE' not in os.environ:
-    print("WARNING: PREDS_FILE not found in environment variables!")
+    logger.error("PREDS_FILE not found in environment variables!")
 
 LABELS_FILE = os.environ.get('LABELS_FILE')
 PREDS_FILE = os.environ.get('PREDS_FILE')
@@ -50,8 +55,7 @@ def get_data():
                 labelled = False
     except FileNotFoundError:
         labelled = False
-        print("Labels file not found. Initialise some labels to create "
-            "the file.")
+        logger.warning("Labels file not found. Initialise some labels to create the file.")
 
     # Load predictions data
     predictions = dict()
@@ -59,8 +63,7 @@ def get_data():
         with open(PREDS_FILE) as f:
             predictions = json.load(f)
     except FileNotFoundError:
-        print("Predictions file not found. Initialise some labels to create "
-            "the file.")
+        logger.warning("Predictions file not found. Initialise some labels to create the file.")
 
     # Retrieve raw text data for each file
     files = file_handler.load_files()
@@ -133,8 +136,11 @@ def reset_data():
     :returns: Empty 204 response object
     """
     # Remove json data files
-    os.remove(LABELS_FILE)
-    os.remove(PREDS_FILE)
+    if os.path.exists(LABELS_FILE):
+        os.remove(LABELS_FILE)
+
+    if os.path.exists(PREDS_FILE):
+        os.remove(PREDS_FILE)
 
     return ('', 204)
